split into reductions test dir
