+++ pandas/tests/reductions/__init__.py
@@ -1,0 +1,4 @@
+"""
+Tests for reductions where we want to test for matching behavior across
+Array, Index, Series, and DataFrame methods.
+"""
